Create MF & Direction API
This was the final project for the VetsInTech Python Fundamentals Course. The premise of the project is that I would act in place of an investment group, looking to to plan a trip where they would tour target properties in two major US cities. My job was to pull bulk data from a real estate API to find Multifamily properties currently on the market, and load the information into an excel file. Once that was done, I would filter the properties to meet the group's investment criteria, and then use a separate directions API to gather the distances from the airport to the target property.